refactor(domain): Money의 int money > long money 형 변환
피드백주신대로, 복권 구매 시 사용자가 입력하는 금액이나 복권 총 당첨금액이 int형을 넘어서는 경우를 대비해 long 형으로 변경

Money
- int money > long money

LongUtils
- Money의 money를 long으로 변경함에 따른 형변환 메소드 추가

+++ src/main/java/lotto/domain/Lottos.java
@@ -28,7 +28,7 @@ public class Lottos {
     }
 
     public Money findTotalPrice() {
-        return Money.createMoney(lottos.size() * LOTTO_PRICE);
+        return Money.createMoney((long) lottos.size() * LOTTO_PRICE);
     }
 
     public LottoResults createLottoResults(Lotto winningLotto) {
